+++ web-ui/src/components/root-components/constants.ts
@@ -1,4 +1,22 @@
 export const DRAWER_WIDTH = 240;
 
-export type AppPage = 'dashboard' | 'blobs' | 'factory' | 'standings' | 'calendar' | 'hall-of-fame' | 'results' | 'event' | 'login';
-export const MENU_ITEMS: AppPage[] = ['dashboard', 'blobs', 'factory', 'standings', 'calendar', 'hall-of-fame', 'results'];
+export type AppPage =
+  | 'dashboard'
+  | 'blobs'
+  | 'factory'
+  | 'standings'
+  | 'calendar'
+  | 'hall-of-fame'
+  | 'results'
+  | 'event'
+  | 'replay'
+  | 'login';
+export const MENU_ITEMS: AppPage[] = [
+  'dashboard',
+  'blobs',
+  'factory',
+  'standings',
+  'calendar',
+  'hall-of-fame',
+  'results',
+];